refac(zk): move `VerifyingEvaluator` to `zk::plonk` namespace
