+++ tachyon/zk/lookup/halo2/BUILD.bazel
@@ -89,8 +89,8 @@ tachyon_cc_library(
         "//tachyon/crypto/commitments:polynomial_openings",
         "//tachyon/zk/lookup:argument",
         "//tachyon/zk/lookup:verifier",
-        "//tachyon/zk/lookup:verifying_evaluator",
         "//tachyon/zk/plonk/base:l_values",
+        "//tachyon/zk/plonk/expressions:verifying_evaluator",
         "//tachyon/zk/plonk/halo2:proof",
     ],
 )
